+++ cli/src/main.rs
@@ -74,6 +74,7 @@ async fn run_impl(args: Args) -> Result<()> {
   let Config {
     accounts,
     recipients,
+    pgp_keybox,
     filters,
     transfer_encoding,
   } = config;
@@ -107,6 +108,7 @@ async fn run_impl(args: Args) -> Result<()> {
     .context("failed to apply filters to message")?;
   let subject = subject.as_deref().unwrap_or("");
   let opts = EmailOpts {
+    pgp_keybox: pgp_keybox.as_deref(),
     transfer_encoding: transfer_encoding.as_deref(),
     ..Default::default()
   };