+++ src/lib.rs
@@ -2,15 +2,21 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 #![allow(
+  clippy::collapsible_else_if,
   clippy::collapsible_if,
   clippy::fn_to_numeric_cast,
   clippy::let_and_return,
   clippy::let_unit_value
 )]
+#![cfg_attr(docsrs, feature(doc_cfg))]
 
 mod config;
+#[cfg(feature = "pgp")]
+mod pgp;
 mod rand;
 
+use std::marker::PhantomData;
+use std::path::Path;
 use std::str;
 use std::str::FromStr as _;
 
@@ -19,9 +25,12 @@ use anyhow::Context as _;
 use anyhow::Error;
 use anyhow::Result;
 
+use lettre::message::header::ContentDisposition;
 use lettre::message::header::ContentTransferEncoding;
 use lettre::message::header::ContentType;
 use lettre::message::MaybeString;
+use lettre::message::MultiPart;
+use lettre::message::SinglePart;
 use lettre::transport::smtp::authentication::Credentials;
 use lettre::AsyncSmtpTransport;
 use lettre::AsyncTransport;
@@ -31,6 +40,8 @@ use lettre::Tokio1Executor;
 pub use crate::config::Account;
 pub use crate::config::SmtpMode;
 
+#[cfg(feature = "pgp")]
+use crate::pgp::encrypt;
 use crate::rand::RandExt as _;
 use crate::rand::Rng;
 
@@ -38,6 +49,16 @@ use crate::rand::Rng;
 /// A type capturing options for capturing a screenshot.
 #[derive(Clone, Debug, Default)]
 pub struct EmailOpts<'input> {
+  /// PGP encrypt the email using the provided keybox file.
+  ///
+  /// The referenced keybox needs to contain public keys for all
+  /// provided recipients or a runtime error will be reported.
+  ///
+  /// With GnuPG, a keybox can be created via `gpg --armor --export
+  /// 'deso@posteo.net' > keybox.gpg`, for example.
+  #[cfg(feature = "pgp")]
+  #[cfg_attr(docsrs, doc(cfg(feature = "pgp")))]
+  pub pgp_keybox: Option<&'input Path>,
   /// The transfer encoding to use.
   ///
   /// This setting is mostly provided as an escape hatch of sorts. If
@@ -46,7 +67,17 @@ pub struct EmailOpts<'input> {
   pub transfer_encoding: Option<&'input str>,
   /// The type is non-exhaustive and open to extension.
   #[doc(hidden)]
-  pub _non_exhaustive: (),
+  pub _phantom: PhantomData<&'input ()>,
+}
+
+
+#[cfg(not(feature = "pgp"))]
+fn encrypt<R, S>(_message: &[u8], _keybox: &Path, _recipients: R) -> Result<Vec<u8>>
+where
+  R: IntoIterator<Item = S>,
+  S: AsRef<str>,
+{
+  unreachable!()
 }
 
 
@@ -73,15 +104,17 @@ where
     })
     .transpose()?
     .unwrap_or(ContentType::TEXT_PLAIN);
-  let email = Message::builder()
-    .from(from)
-    .subject(subject)
-    .header(content_type);
+  let email = Message::builder().from(from).subject(subject);
 
   let EmailOpts {
+    #[cfg(feature = "pgp")]
+    pgp_keybox,
     transfer_encoding,
-    _non_exhaustive: (),
+    _phantom: PhantomData,
   } = opts;
+
+  #[cfg(not(feature = "pgp"))]
+  let pgp_keybox = None;
 
   // We only set the transfer encoding if the user provided one. The
   // reason being that:
@@ -96,7 +129,7 @@ where
     email
   };
 
-  for recipient in recipients {
+  for recipient in recipients.clone() {
     let recipient = recipient.as_ref();
     let to = recipient
       .parse()
@@ -105,17 +138,6 @@ where
     email = email.to(to);
   }
 
-  // We always try to work with string. The reason being that `lettre`
-  // performs line ending conversion only when the data is passed in
-  // as a string, and some mailers reject emails with bare linefeed
-  // line endings.
-  let body = if let Ok(message) = str::from_utf8(message) {
-    MaybeString::String(message.to_string())
-  } else {
-    MaybeString::Binary(message.to_vec())
-  };
-
-  let email = email.body(body).context("failed to create email message")?;
   let creds = Credentials::new(account.user.to_string(), account.password.to_string());
 
   let mailer = match account.smtp_mode {
@@ -132,6 +154,61 @@ where
       .context("failed to create STARTTLS SMTP mailer")?
       .credentials(creds)
       .build(),
+  };
+
+  let email = if let Some(keybox) = pgp_keybox {
+    let inner = MultiPart::mixed().singlepart(
+      SinglePart::builder()
+        .header(content_type)
+        .body(message.to_vec()),
+    );
+
+    // TODO: Ideally we'd also sign the message, but that's a different
+    //       pandora's box and not as important at this point.
+    let message =
+      encrypt(&inner.formatted(), keybox, recipients).context("failed to encrypt message")?;
+    // We always ASCII armor the message, so we do not expect it to ever
+    // be *not* a valid UTF-8 string.
+    let message =
+      str::from_utf8(&message).context("PGP encrypted message is not a valid UTF-8 string")?;
+
+    let parts = MultiPart::encrypted("application/pgp-encrypted".to_owned())
+      .singlepart(
+        SinglePart::builder()
+          .header(
+            ContentType::parse("application/pgp-encrypted")
+              .context("failed to parse 'application/pgp-encrypted' content type header")?,
+          )
+          .body(String::from("Version: 1")),
+      )
+      .singlepart(
+        SinglePart::builder()
+          .header(
+            ContentType::parse(r#"application/octet-stream; name="encrypted.asc""#)
+              .context("failed to parse 'application/octet-stream' content type header")?,
+          )
+          .header(ContentDisposition::inline_with_name("encrypted.asc"))
+          .body(message.to_string()),
+      );
+
+    email
+      .multipart(parts)
+      .context("failed to create email message")?
+  } else {
+    // We always try to work with string. The reason being that `lettre`
+    // performs line ending conversion only when the data is passed in
+    // as a string, and some mailers reject emails with bare linefeed
+    // line endings.
+    let body = if let Ok(message) = str::from_utf8(message) {
+      MaybeString::String(message.to_string())
+    } else {
+      MaybeString::Binary(message.to_vec())
+    };
+
+    email
+      .header(content_type)
+      .body(body)
+      .context("failed to create email message")?
   };
 
   let _mailer = mailer