+++ cli/CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Added support for PGP encrypting emails
 - Made transfer encoding to use configurable
 - Made content type to use configurable via `--content-type` argument